rename: 디렉토리 구조 변경: views, controllers, util
