+++ public/js/index.js
@@ -1,7 +1,7 @@
-import { initEventListener } from "./event/event-listener.js";
-import { dowTopDefault } from "./blocks/dow-top.js";
-import { renderingDowNavDefault } from "./rendering/nav.js";
-import { renderingMainAd } from "./rendering/main-ad-banner.js";
+import { initEventListener } from "./controllers/event-listener.js";
+import { dowTopDefault } from "./views/blocks/dow-top.js";
+import { renderingDowNavDefault } from "./views/render/nav.js";
+import { renderingMainAd } from "./views/render/main-ad-banner.js";
 
 const renderingDefault = () => {
   renderingMainAd("홈");
