+++ public/js/views/components/event-banner.js
