+++ public/js/views/components/promotion-banner.js
