+++ public/js/views/components/sub-ad-banner.js
