+++ public/js/views/components/webtoon-card.js
