+++ public/js/views/components/webtoon-long-card.js
