+++ public/js/views/render/nav.js
@@ -1,4 +1,4 @@
-import { getToday } from "../utils.js";
+import { getToday } from "../../util/utils.js";
 
 const renderingNav = (selectedNav, selectedElement) => {
   document.querySelectorAll(`.${selectedNav} li`).forEach((e) => {
